requirements.txt
streamlit==0.90.0
Pillow==8.3.1
pandas==1.3.3
plotly==5.3.1
matplotlib==3.4.3
geopandas==0.10.2